+++ src/query/typed_value.hpp
@@ -8,6 +8,7 @@
 #include <unordered_set>
 #include <vector>
 
+#include "query/path.hpp"
 #include "storage/edge_accessor.hpp"
 #include "storage/property_value.hpp"
 #include "storage/vertex_accessor.hpp"
@@ -16,8 +17,6 @@
 #include "utils/total_ordering.hpp"
 
 namespace query {
-
-typedef traversal_template::Path<VertexAccessor, EdgeAccessor> Path;
 
 // TODO: Neo4j does overflow checking. Should we also implement it?
 /**